+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix jakarta.inject annotations requirement (NoClassDefFoundError) (#430)
+
 ### 6.3.1 (2025-05-12)
 * Fix NoClassDefFoundError on guicey startup due to junit classes leak into core (#428)